+++ typescript/sdk/src/consts/multisigIsm.ts
@@ -3,6 +3,15 @@ import { ChainMap } from '../types.js';
 
 // TODO: consider migrating these to the registry too
 export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
+  alephzeroevm: {
+    threshold: 2,
+    validators: [
+      '0xcae8fab142adc4e434bb7409e40dd932cc3851aa',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
   alfajores: {
     threshold: 2,
     validators: [
@@ -35,6 +44,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   arbitrumsepolia: {
     threshold: 1,
     validators: ['0x09fabfbca0b8bf042e2a1161ee5010d147b0f603'],
+  },
+
+  arcadiatestnet: {
+    threshold: 1,
+    validators: ['0x7ce5973d3f22971546efb86f5a0417c1248e92f5'],
   },
 
   astar: {
@@ -78,6 +92,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   basesepolia: {
     threshold: 1,
     validators: ['0x82e3b437a2944e3ff00258c93e72cd1ba5e0e921'],
+  },
+
+  berabartio: {
+    threshold: 1,
+    validators: ['0x541dd3cb282cf869d72883557badae245b63e1fd'],
   },
 
   bitlayer: {
@@ -126,6 +145,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  camptestnet: {
+    threshold: 1,
+    validators: ['0x238f40f055a7ff697ea6dbff3ae943c9eae7a38e'],
+  },
+
   celo: {
     threshold: 3,
     validators: [
@@ -152,6 +176,20 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x0874967a145d70b799ebe9ed861ab7c93faef95a',
       '0xd767ea1206b8295d7e1267ddd00e56d34f278db6',
     ],
+  },
+
+  chiliz: {
+    threshold: 2,
+    validators: [
+      '0x82d024f453b1a3f3f6606226f06b038da27596f3',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
+  citreatestnet: {
+    threshold: 1,
+    validators: ['0x60d7380a41eb95c49be18f141efd2fde5e3dba20'],
   },
 
   connextsepolia: {
@@ -255,6 +293,20 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  flow: {
+    threshold: 2,
+    validators: [
+      '0x3aee1090318e9c54d1d23194dcd0f2bee00ddc97',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
+  formtestnet: {
+    threshold: 1,
+    validators: ['0x72ad7fddf16d17ff902d788441151982fa31a7bc'],
+  },
+
   fraxtal: {
     threshold: 2,
     validators: [
@@ -295,6 +347,20 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   holesky: {
     threshold: 1,
     validators: ['0x7ab28ad88bb45867137ea823af88e2cb02359c03'], // TODO
+  },
+
+  hyperliquidevmtestnet: {
+    threshold: 1,
+    validators: ['0xea673a92a23ca319b9d85cc16b248645cd5158da'],
+  },
+
+  immutablezkevm: {
+    threshold: 2,
+    validators: [
+      '0xa787c2952a4d22f776ee6e87e828e6f75de24330',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
   },
 
   inevm: {
@@ -351,6 +417,15 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  lumia: {
+    threshold: 2,
+    validators: [
+      '0x9e283254ed2cd2c80f007348c2822fc8e5c2fa5f',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
   mantapacific: {
     threshold: 5,
     validators: [
@@ -378,6 +453,15 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: [
       '0xc1d6600cb9326ed2198cc8c4ba8d6668e8671247',
       '0xcf0211fafbb91fd9d06d7e306b30032dc3a1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
+  metall2: {
+    threshold: 2,
+    validators: [
+      '0x1b000e1e1f0a032ed382c6d69a2d58f6fe773c09',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
       '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
     ],
   },
@@ -442,6 +526,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  odysseytestnet: {
+    threshold: 1,
+    validators: ['0xcc0a6e2d6aa8560b45b384ced7aa049870b66ea3'],
+  },
+
   oortmainnet: {
     threshold: 2,
     validators: [
@@ -501,11 +590,29 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  polynomial: {
+    threshold: 2,
+    validators: [
+      '0xa63ad0891e921ad5947d57e05831fabb9816eca7',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
   proofofplay: {
     threshold: 2,
     validators: [
       '0xcda40baa71970a06e5f55e306474de5ca4e21c3b',
       '0xcf0211fafbb91fd9d06d7e306b30032dc3a1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
+  rari: {
+    threshold: 2,
+    validators: [
+      '0x989d6862e09de21337078efbd86843a3eb1133e3',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
       '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
     ],
   },
@@ -524,6 +631,15 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: [
       '0x1400b9737007f7978d8b4bbafb4a69c83f0641a7',
       '0xcf0211fafbb91fd9d06d7e306b30032dc3a1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
+    ],
+  },
+
+  rootstock: {
+    threshold: 2,
+    validators: [
+      '0xcb8e3a72cf427feff27416d0e2ec375a052eaaee',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
       '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
     ],
   },
@@ -609,6 +725,16 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: ['0xd4ce8fa138d4e083fc0e480cca0dbfa4f5f30bd5'],
   },
 
+  soneiumtestnet: {
+    threshold: 1,
+    validators: ['0x2e2101020ccdbe76aeda1c27823b0150f43d0c63'],
+  },
+
+  sonictestnet: {
+    threshold: 1,
+    validators: ['0x62e6591d00daec3fb658c3d19403828b4e9ddbb3'],
+  },
+
   stride: {
     threshold: 6,
     validators: [
@@ -622,6 +748,20 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0xf54982134e52Eb7253236943FBffE0886C5bde0C', // L5
       '0x5937b7cE1029C3Ec4bD8e1AaCc0C0f9422654D7d', // stakecito
       '0x3a446ed2923c08445af06e53f0acb558c0e0413c', // staked
+    ],
+  },
+
+  suavetoliman: {
+    threshold: 1,
+    validators: ['0xf58f6e30aabba34e8dd7f79b3168507192e2cc9b'],
+  },
+
+  superposition: {
+    threshold: 2,
+    validators: [
+      '0x5978d0e6afa9270ddb87cff43a8fa7a763a5dfc4',
+      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
+      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
     ],
   },
 
@@ -647,6 +787,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0xcf0211fafbb91fd9d06d7e306b30032dc3a1934f', // merkly
       '0xe271ef9a6e312540f099a378865432fa73f26689', // tangle
     ],
+  },
+
+  unichaintestnet: {
+    threshold: 1,
+    validators: ['0x5e99961cf71918308c3b17ef21b5f515a4f86fe5'],
   },
 
   viction: {