+++ typescript/sdk/src/metadata/chainMetadataTypes.ts
@@ -4,7 +4,9 @@
  */
 import { SafeParseReturnType, z } from 'zod';
 
-import { ProtocolType } from '@hyperlane-xyz/utils';
+import { ProtocolType, objMerge } from '@hyperlane-xyz/utils';
+
+import { ChainMap } from '../types.js';
 
 import { ZChainName, ZNzUint, ZUint } from './customZodTypes.js';
 
@@ -17,6 +19,10 @@ export enum ExplorerFamily {
 
 export enum ChainTechnicalStack {
   ArbitrumNitro = 'arbitrumnitro',
+  OpStack = 'opstack',
+  PolygonCDK = 'polygoncdk',
+  PolkadotSubstrate = 'polkadotsubstrate',
+  ZkSync = 'zksync',
   Other = 'other',
 }
 
@@ -67,6 +73,29 @@ export const RpcUrlSchema = z.object({
 
 export type RpcUrl = z.infer<typeof RpcUrlSchema>;
 
+export const BlockExplorerSchema = z.object({
+  name: z.string().describe('A human readable name for the explorer.'),
+  url: z.string().url().describe('The base URL for the explorer.'),
+  apiUrl: z
+    .string()
+    .url()
+    .describe('The base URL for requests to the explorer API.'),
+  apiKey: z
+    .string()
+    .optional()
+    .describe(
+      'An API key for the explorer (recommended for better reliability).',
+    ),
+  family: z
+    .nativeEnum(ExplorerFamily)
+    .optional()
+    .describe(
+      'The type of the block explorer. See ExplorerFamily for valid values.',
+    ),
+});
+
+export type BlockExplorer = z.infer<typeof BlockExplorerSchema>;
+
 export const NativeTokenSchema = z.object({
   name: z.string(),
   symbol: z.string(),
@@ -87,28 +116,7 @@ export const ChainMetadataSchemaObject = z.object({
     .describe('The human readable address prefix for the chains using bech32.'),
 
   blockExplorers: z
-    .array(
-      z.object({
-        name: z.string().describe('A human readable name for the explorer.'),
-        url: z.string().url().describe('The base URL for the explorer.'),
-        apiUrl: z
-          .string()
-          .url()
-          .describe('The base URL for requests to the explorer API.'),
-        apiKey: z
-          .string()
-          .optional()
-          .describe(
-            'An API key for the explorer (recommended for better reliability).',
-          ),
-        family: z
-          .nativeEnum(ExplorerFamily)
-          .optional()
-          .describe(
-            'The type of the block explorer. See ExplorerFamily for valid values.',
-          ),
-      }),
-    )
+    .array(BlockExplorerSchema)
     .optional()
     .describe('A list of block explorers with data for this chain'),
 
@@ -230,7 +238,7 @@ export const ChainMetadataSchemaObject = z.object({
 
   rpcUrls: z
     .array(RpcUrlSchema)
-    .nonempty()
+    .min(1)
     .describe('The list of RPC endpoints for interacting with the chain.'),
 
   slip44: z.number().optional().describe('The SLIP-0044 coin type.'),
@@ -341,11 +349,6 @@ export type ChainMetadata<Ext = object> = z.infer<
 > &
   Ext;
 
-export type BlockExplorer = Exclude<
-  ChainMetadata['blockExplorers'],
-  undefined
->[number];
-
 export function safeParseChainMetadata(
   c: ChainMetadata,
 ): SafeParseReturnType<ChainMetadata, ChainMetadata> {
@@ -372,4 +375,18 @@ export function getReorgPeriod(chainMetadata: ChainMetadata): number {
   if (chainMetadata.blocks?.reorgPeriod !== undefined)
     return chainMetadata.blocks.reorgPeriod;
   else throw new Error('Chain has no reorg period');
+}
+
+export function mergeChainMetadata(
+  base: ChainMetadata,
+  overrides: Partial<ChainMetadata> | undefined,
+): ChainMetadata {
+  return objMerge<ChainMetadata>(base, overrides || {}, 10, true);
+}
+
+export function mergeChainMetadataMap(
+  base: ChainMap<ChainMetadata>,
+  overrides: ChainMap<Partial<ChainMetadata> | undefined> | undefined,
+): ChainMap<ChainMetadata> {
+  return objMerge<ChainMap<ChainMetadata>>(base, overrides || {}, 10, true);
 }